Add a column with names of offshore regions

Add a column with names of offshore regions

Usage

add_reg_off(x, regN, regN_off = paste0(regN, "_off"), offshore = "offshore")

Arguments

x

data frame with required two columns: offshore and regN

regN

character name of the column with region names

regN_off

character name of the column with offshore region names, that will be added to the data frame, default is {regN}_off

offshore

character name of the column with offshore indicator, default is 'offshore'

Value

data frame with added column regN_off