Load sf (simple feature) object of solar or wind clusters

Load sf (simple feature) object of solar or wind clusters

Usage

get_ideea_cl_sf(
  resource = "sol",
  nreg = 5,
  tol = 0.05,
  data = "merra2",
  clusters_sf_file = "locid_{resource}_cl_r{nreg}_sf.RData"
)

Arguments

resource

character, name of the resource: "sol" or "win"

nreg

integer, the number of regions

tol

numeric, the tolerance of the information loss in the clustering from 0 to 1. If 0 the number of clusters is equal to the number of locations in MERRA2 database. When tol is 1, there is only one cluster per region.

data

character, the name of the data source, the directory in the ideea_extra directory. Currently only "merra2" is supported.

clusters_sf_file

character, the name of the file with the clusters in the data directory. Default is "locid_resource_cl_rnreg_sf.RData", available in ideea_data.

Value

sf, the simple feature object with the clusters for the requested resource.

Examples

# don't run
get_ideea_cl_sf(resource = "sol", nreg = 5)
#> Error in str_detect(names(x), pattern = "cf_"): could not find function "str_detect"
get_ideea_cl_sf(resource = "win", nreg = 32, tol = 0.01)
#> Error in str_detect(names(x), pattern = "cf_"): could not find function "str_detect"